% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_0_1_talk_initialize.R
\name{find_talkrpp_env}
\alias{find_talkrpp_env}
\title{Find talk required python packages env}
\usage{
find_talkrpp_env()
}
\description{
check whether conda/virtual environment for talk required python pacakges exists
}
\keyword{internal}
