% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_1_talkEmbed.R
\name{talkEmbed}
\alias{talkEmbed}
\title{Transform audio recordings to embeddings}
\usage{
talkEmbed(talk, model = "WhiSPA")
}
\arguments{
\item{talk}{(audio file) .}

\item{model}{(character) Path or name of the model to use for embedding the audio data.}
}
\value{
A tibble with embeddings.
}
\description{
Transform audio recordings to embeddings
}
\examples{
# Transform audio recordings in the example dataset:
# voice_data (included in talk-package), to embeddings.
\dontrun{

}

}
\seealso{
\code{\link{talkText}}.
}
