% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_0_1_talk_initialize.R
\name{talkrpp_initialize}
\alias{talkrpp_initialize}
\title{Initialize talk required python packages}
\usage{
talkrpp_initialize(
  python_executable = NULL,
  virtualenv = NULL,
  condaenv = "talkrpp_condaenv",
  ask = FALSE,
  refresh_settings = FALSE,
  save_profile = FALSE,
  check_env = TRUE,
  talkEmbed_test = FALSE,
  prompt = TRUE
)
}
\arguments{
\item{python_executable}{the full path to the Python executable, for which
talk required python packages is installed.}

\item{virtualenv}{set a path to the Python virtual environment with talk required python packages
installed Example: \code{virtualenv = "~/myenv"}}

\item{condaenv}{set a path to the anaconda virtual environment with talk required python packages
installed Example: \code{condalenv = "myenv"}}

\item{ask}{logical; if \code{FALSE}, use the first talk required python packages installation found;
if \code{TRUE}, list available talk required python packages installations and prompt the user for
which to use. If another (e.g. \code{python_executable}) is set, then this
value will always be treated as \code{FALSE}.}

\item{refresh_settings}{logical; if \code{TRUE}, talk will ignore the saved
settings in the profile and initiate a search of new settings.}

\item{save_profile}{logical; if \code{TRUE}, the current talk required python packages setting will
be saved for the future use.}

\item{check_env}{logical; check whether conda/virtual environment generated
by \code{talkrpp_install()} exists}

\item{talkEmbed_test}{logical; Test whether function (talkEmbed) that requires python packages works.}

\item{prompt}{logical; asking whether user wants to set the environment as default.}
}
\description{
Initialize talk required python packages to call from R.
}
