% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_0_talk_install.R
\name{talkrpp_uninstall}
\alias{talkrpp_uninstall}
\title{Uninstall talkrpp conda environment}
\usage{
talkrpp_uninstall(conda = "auto", prompt = TRUE, envname = "talkrpp_condaenv")
}
\arguments{
\item{conda}{path to conda executable, default to "auto" which automatically
finds the path}

\item{prompt}{logical; ask whether to proceed during the installation}

\item{envname}{character; name of conda environment to remove}
}
\description{
Removes the conda environment created by talkrpp_install()
}
